sealed class serialisation refactored
